Add support for fractional screenscaling with layers
Thanks to Hermann Vogt

+++ IGraphics/Drawing/IGraphicsCanvas.cpp
@@ -389,7 +389,7 @@ APIBitmap* IGraphicsCanvas::LoadAPIBitmap(const char* name, const void* pData, i
   return new Bitmap(canvas, name, scale);
 }
 
-APIBitmap* IGraphicsCanvas::CreateAPIBitmap(int width, int height, int scale, double drawScale, bool cacheable)
+APIBitmap* IGraphicsCanvas::CreateAPIBitmap(int width, int height, float scale, double drawScale, bool cacheable)
 {
   return new Bitmap(width, height, scale, drawScale);
 }
